+++ docs/change-log.md
@@ -1,11 +1,20 @@
+## 2.1.2 (October 23, 2016)
+
+- patch `dio.renderToString` className should map to class
+- patch `dio.Children.only`
+- improve compatibility with commonJS enviroments
+- add `dio.unmountComponentAtNode` top level api, mirrors `React`
+- add `dio.applyMiddleware`, `dio.combineReducers` and `dio.compose` to top level api
+
 ## 2.1.1 (October 23, 2016)
 
-- add dio.Children top level api, mirros React.Children
+- add `dio.Children` top level api, mirrors `React`
 - better intergration with webpack
 - patch `dio.createElement` rename conflict
 - patch `dio.createElement` assign children components
 - patch assignment of component `this.props.children`
 - patch issue with es6 Components class `.defaultProps` not registering
+- pathc redux store initialState
 
 ## 2.1.0 (October 18, 2016)